# kraken: optimization of direct path