* For direct path of 10 to 20minutes in a query the processing time is reduced from 14ms to 6 ms.
* For direct path of 11 to 22 minutes in another query the processing time is reduced from "slow request" (> 1000 ms) to  between 42ms and 217ms  

*concerned ticket: http://jira.canaltp.fr/browse/ITI-373